refactor: separate compilation logic from GitManager
Extract compilation operations into dedicated CompilationManager to improve separation of concerns and follow single responsibility principle.

Changes:
- Create CompilationManager for handling reth and reth-bench compilation
- Move compile_reth() and compile_reth_bench() from GitManager to CompilationManager
- Update GitManager to focus solely on git operations (branching, validation, etc.)
- Update CLI workflow to use separate managers for git and compilation operations
- Remove unused imports from git.rs

Benefits:
- Better separation of concerns between git and build operations
- More maintainable and focused class responsibilities
- Easier to modify compilation logic independently of git operations
